Cache properties as late as possible if equal cardinality ()
Cache properties as late as possible.

- Cache properties as late as possible if separate steps have equal 
  cardinality. While there are cases when this can potentially lead to 
  worse planning, we argue that we should not optimise for such cases since 
  they build on the assumption that the cardinality will be incorrect.
- Introduce threshold for how small cardinality changes we care about in push 
  down property reads.
- Add skip to the cardinality model in a similar way that we do for limit.

+++ community/cypher/cypher-planner/src/test/scala/org/neo4j/cypher/internal/compiler/helpers/LogicalPlanBuilder.scala
@@ -62,7 +62,7 @@ class LogicalPlanBuilder(wholePlan: Boolean = true, resolver: Resolver = new Log
 
   def getSemanticTable: SemanticTable = semanticTable
 
-  def withCardinality(x: Int): LogicalPlanBuilder = {
+  def withCardinality(x: Double): LogicalPlanBuilder = {
     cardinalities.set(idOfLastPlan, Cardinality(x))
     this
   }
